+++ peptdeep/pipeline_api.py
@@ -14,7 +14,7 @@ from alphabase.peptide.fragment import (
     concat_precursor_fragment_dataframes
 )
 from alphabase.spectral_library.reader import (
-    SWATHLibraryReader
+    LibraryReaderBase
 )
 
 from peptdeep.spec_lib.translate import mod_to_unimod_dict
@@ -23,18 +23,24 @@ from peptdeep.utils import (
     logging, set_logger, 
     show_platform_info, show_python_info
 )
-from peptdeep.rescore.percolator import Percolator
+# from peptdeep.rescore.percolator import Percolator
 from peptdeep.spec_lib.library_factory import (
     library_maker_provider
 )
 
 from peptdeep.pretrained_models import ModelManager
 
-from peptdeep.rescore.feature_extractor import match_one_raw
+# from peptdeep.rescore.feature_extractor import match_one_raw
+from alpharaw.match.psm_match import PepSpecMatch, PepSpecMatch_DIA
 
-from peptdeep.utils import parse_ms_file_names_to_dict
+from peptdeep.model.ms2 import calc_ms2_similarity
 
-from peptdeep.utils import process_bar
+def _check_is_file(fname):
+    if isinstance(fname, str) and not os.path.isfile(fname):
+        logging.info(f" -- File `{fname}` does not exist.")
+        return False
+    else:
+        return True
 
 def import_psm_df(psm_files:list, psm_type:str)->pd.DataFrame:
     """Import PSM files of a search engine as a pd.DataFrame
@@ -56,15 +62,49 @@ def import_psm_df(psm_files:list, psm_type:str)->pd.DataFrame:
         psm_type, 
         modificatin_mapping=global_settings[
             'model_mgr']['transfer'
-        ]['other_modification_mapping']
+        ]['psm_modification_mapping']
     )
     psm_df_list = []
     for psm_file in psm_files:
-        if not os.path.isfile(psm_file): continue
+        if not _check_is_file(psm_file): continue
         psm_reader.import_file(psm_file)
 
         psm_df_list.append(psm_reader.psm_df)
     return pd.concat(psm_df_list).reset_index(drop=True)
+
+def get_median_pccs_for_dia_psms(
+    psm_match:PepSpecMatch_DIA,
+    psm_df:pd.DataFrame, 
+    fragment_mz_df:pd.DataFrame,
+    fragment_intensity_df:pd.DataFrame,
+):
+    _frag_df = fragment_intensity_df.mask(
+        fragment_mz_df<psm_match.min_frag_mz, 0.0
+    )
+    frag_len = len(_frag_df)//psm_match.max_spec_per_query
+    psm_len = len(psm_match.psm_df)//psm_match.max_spec_per_query
+    _df = psm_match.psm_df.iloc[:psm_len].copy()
+    median_pccs = np.zeros(len(psm_df))
+    metrics_list = []
+    for i in range(psm_match.max_spec_per_query):
+        pcc_list = []
+        for j in range(psm_match.max_spec_per_query):
+            if i == j: continue
+            _df,metrics_df = calc_ms2_similarity(
+                _df,
+                _frag_df[i*frag_len:(i+1)*frag_len],
+                _frag_df[j*frag_len:(j+1)*frag_len],
+            )
+            pcc_list.append(_df.PCC.values)
+            metrics_list.append(metrics_df)
+        pccs = np.median(np.array(pcc_list),axis=0)
+        median_pccs[i*psm_len:(i+1)*psm_len] = pccs
+
+    logging.info(
+        f"Average MS2 similarity metrics among {psm_match.max_spec_per_query} DIA scans at frag_mz>={psm_match.min_frag_mz}:\n" 
+        f"{str(sum(metrics_list)/len(metrics_list))}"
+    )
+    return median_pccs
 
 def match_psms()->Tuple[pd.DataFrame,pd.DataFrame]:
     """
@@ -96,6 +136,8 @@ def match_psms()->Tuple[pd.DataFrame,pd.DataFrame]:
         for _type in global_settings['model']['frag_types']:
             if 'modloss' not in _type:
                 frag_types.append(_type)
+    else:
+        frag_types = global_settings['model']['frag_types']
 
     max_charge = global_settings['model']['max_frag_charge']
     charged_frag_types = get_charged_frag_types(frag_types, max_charge)
@@ -105,35 +147,58 @@ def match_psms()->Tuple[pd.DataFrame,pd.DataFrame]:
         mgr_settings['transfer']['psm_type'],
     )
 
-    ms2_file_dict = parse_ms_file_names_to_dict(
-        mgr_settings['transfer']['ms_files']
-    )
-    
-    psm_df_list = []
-    matched_intensity_df_list = []
-    df_groupby_raw = psm_df.groupby('raw_name')
-    for raw_name, df in process_bar(
-        df_groupby_raw, df_groupby_raw.ngroups
-    ):
-        if raw_name not in ms2_file_dict:
-            continue
-        (
-            df, _, inten_df, _
-        ) = match_one_raw(
-            df, ms2_file_dict[raw_name],
-            mgr_settings['transfer']['ms_file_type'],
-            charged_frag_types,
-            global_settings['peak_matching']['ms2_ppm'], 
-            global_settings['peak_matching']['ms2_tol_value'],
-            calibrate_frag_mass_error=False,
-        )
-        psm_df_list.append(df)
-        matched_intensity_df_list.append(inten_df)
+    ms2_file_list = mgr_settings['transfer']['ms_files']
+    for _ms_file in ms2_file_list:
+        if not os.path.isfile(_ms_file):
+            logging.warn(f"`{_ms_file}` is invalid, please check the paths of `ms_files`")
 
-    return concat_precursor_fragment_dataframes(
-        psm_df_list,
-        matched_intensity_df_list
+    if (
+        mgr_settings['transfer']['psm_type'].lower()
+        in mgr_settings['transfer']['dda_psm_types'] 
+    ):
+        psm_match = PepSpecMatch(
+            charged_frag_types=charged_frag_types,
+            use_ppm=global_settings['peak_matching']['ms2_ppm'], 
+            tol_value=global_settings['peak_matching']['ms2_tol_value'],
+        )
+    else:
+        psm_match = PepSpecMatch_DIA(
+            charged_frag_types=charged_frag_types,
+            use_ppm=global_settings['peak_matching']['ms2_ppm'], 
+            tol_value=global_settings['peak_matching']['ms2_tol_value'],
+        )
+        psm_match.max_spec_per_query = 5
+        psm_match.min_frag_mz = global_settings["library"]["output_tsv"]["min_fragment_mz"]
+
+    thread_num = global_settings["thread_num"]
+    if len(ms2_file_list) > thread_num:
+        psm_match.ms_loader_thread_num = 1
+    else:
+        psm_match.ms_loader_thread_num = thread_num//len(ms2_file_list)
+        thread_num = len(ms2_file_list)
+
+    (
+        psm_df, frag_mz_df,
+        frag_inten_df, frag_mz_err_df,
+    ) = psm_match.match_ms2_multi_raw(
+        psm_df=psm_df,
+        ms_files=ms2_file_list,
+        ms_file_type=mgr_settings['transfer']['ms_file_type'],
+        process_num=thread_num,
     )
+
+    logging.info(f"Extracted {len(psm_df)} PSMs.")
+
+    if isinstance(psm_match, PepSpecMatch_DIA):
+        psm_df["median_pcc"] = get_median_pccs_for_dia_psms(
+            psm_match, psm_df,
+            frag_mz_df, frag_inten_df
+        )
+
+        psm_df = psm_df.query("score>=6 and median_pcc>=0.9")
+        logging.info(f"Kept {len(psm_df)} PSMs at ion_count>=6 and median_pcc>=0.9 for training/testing.")
+
+    return psm_df, frag_inten_df
 
 def transfer_learn(verbose=True):
     """Transfer learn / refine the RT/CCS(/MS2) models.
@@ -191,6 +256,7 @@ def transfer_learn(verbose=True):
             log_level=global_settings['log_level'],
             overwrite=True, stream=True, 
         )
+        logging.info("[PeptDeep] Running train task ...")
         show_platform_info()
         show_python_info()
 
@@ -198,35 +264,36 @@ def transfer_learn(verbose=True):
         model_mgr.reset_by_global_settings()
 
         logging.info('Loading PSMs and extracting fragments ...')
+        
         if (
-            model_mgr.psm_num_to_train_ms2 > 0 and 
-            len(mgr_settings['transfer']['ms_files'])>0
+            mgr_settings['transfer']['psm_type'].lower() == 'speclib_tsv'
         ):
-                psm_df, frag_df = match_psms()
+            dfs = []
+            frag_inten_dfs = []
+            for psm_file in mgr_settings['transfer']['psm_files']:
+                _lib = LibraryReaderBase()
+                if not _check_is_file(psm_file): continue
+                dfs.append(_lib.import_file(psm_file))
+                frag_inten_dfs.append(_lib.fragment_intensity_df)
+            psm_df, frag_df = concat_precursor_fragment_dataframes(
+                dfs, frag_inten_dfs
+            )
+        elif len(mgr_settings['transfer']['ms_files'])>0:
+            psm_df, frag_df = match_psms()
         else:
-            if (
-                mgr_settings['transfer']['ms_file_type'].lower()== 'speclib_tsv'  
-                and mgr_settings['transfer']['psm_type'].lower() == 'speclib_tsv'
-            ):
-                dfs = []
-                frag_inten_dfs = []
-                for psm_file in mgr_settings['transfer']['psm_files']:
-                    _lib = SWATHLibraryReader()
-                    dfs.append(_lib.import_file(psm_file))
-                    frag_inten_dfs.append(_lib.fragment_intensity_df)
-                psm_df, frag_df = concat_precursor_fragment_dataframes(
-                    dfs, frag_inten_dfs
-                )
-            else:
-                psm_df = import_psm_df(
-                    mgr_settings['transfer']['psm_files'],
-                    mgr_settings['transfer']['psm_type'],
-                )
-                frag_df = None
+            psm_df = import_psm_df(
+                mgr_settings['transfer']['psm_files'],
+                mgr_settings['transfer']['psm_type'],
+            )
+            frag_df = None
+        
+        if model_mgr.psm_num_to_train_ms2 <= 0:
+            frag_df = None
 
-        logging.info("Training CCS model ...")
-        model_mgr.train_ccs_model(psm_df)
-        logging.info("Finished training CCS model")
+        if "ccs" in psm_df.columns and (psm_df.ccs!=0).all():
+            logging.info("Training CCS model ...")
+            model_mgr.train_ccs_model(psm_df)
+            logging.info("Finished training CCS model")
 
         logging.info("Training RT model ...")
         model_mgr.train_rt_model(psm_df)
@@ -240,24 +307,15 @@ def transfer_learn(verbose=True):
         model_mgr.ccs_model.save(os.path.join(output_folder, 'ccs.pth'))
         model_mgr.rt_model.save(os.path.join(output_folder, 'rt.pth'))
         model_mgr.ms2_model.save(os.path.join(output_folder, 'ms2.pth'))
+        
+        save_yaml(
+            os.path.join(output_folder, 'peptdeep_settings.yaml'),
+            global_settings
+        )
         logging.info(f"Models were saved in {output_folder}")
     except Exception as e:
         logging.error(traceback.format_exc())
         raise e
-
-def _get_delimiter(tsv_file:str):
-    with open(tsv_file, "r") as f:
-        line = f.readline().strip()
-        if '\t' in line: return '\t'
-        elif ',' in line: return ','
-        else: return '\t'
-
-def read_peptide_table(tsv_file:str)->pd.DataFrame:
-    sep = _get_delimiter(tsv_file)
-    df = pd.read_csv(tsv_file, sep=sep, keep_default_na=False)
-    if 'mod_sites' in df.columns:
-        df['mod_sites'] = df.mod_sites.astype('U')
-    return df
 
 def generate_library():
     """Generate/predict a spectral library.
@@ -288,6 +346,8 @@ def generate_library():
             log_level=global_settings['log_level'],
             overwrite=True, stream=True, 
         )
+        logging.info("[PeptDeep] Running library task ...")
+        logging.info(f"Input files ({lib_settings['infile_type']}): " + str(lib_settings['infiles']))
         show_platform_info()
         show_python_info()
 
@@ -298,14 +358,12 @@ def generate_library():
             lib_settings['infile_type'],
             model_manager=model_mgr
         )
-        if lib_settings['infile_type'] == 'fasta':
+
+        if lib_settings['infile_type'].lower() in library_maker_provider.library_maker_dict:
             lib_maker.make_library(lib_settings['infiles'])
-        else:
-            df_list = []
-            for file_path in lib_settings['infiles']:
-                df_list.append(read_peptide_table(file_path))
-            df = pd.concat(df_list, ignore_index=True)
-            lib_maker.make_library(df)
+        else: # PSMReaderLibraryMaker
+            lib_maker.make_library((lib_settings['infile_type'],lib_settings['infiles']))
+        
         save_yaml(
             os.path.join(output_folder, 'peptdeep_settings.yaml'),
             global_settings
@@ -333,72 +391,72 @@ def generate_library():
         logging.error(traceback.format_exc())
         raise e
 
-def rescore():
-    """Generate/predict a spectral library.
+# def rescore():
+#     """Generate/predict a spectral library.
     
-    All required information in global_settings:
+#     All required information in global_settings:
     
-    ```python
-    perc_settings = global_settings['percolator']
-    output_folder = perc_settings['output_folder'] # str. Output folder of the rescored results
-    perc_settings['input_files']['psm_files'] # list of str. all PSM files (at 100% FDR and including decoys) from the search engines
-    perc_settings['input_files']['psm_type'] # str. PSM or search engine type, e.g. pfind, alphapept, maxquant
-    perc_settings['input_files']['ms_file_type'] # str. Could be alphapept_hdf, thermo, ...
-    perc_settings['input_files']['ms_files'] # list of str. MS file list to match MS2 peaks
-    ```
+#     ```python
+#     perc_settings = global_settings['percolator']
+#     output_folder = perc_settings['output_folder'] # str. Output folder of the rescored results
+#     perc_settings['input_files']['psm_files'] # list of str. all PSM files (at 100% FDR and including decoys) from the search engines
+#     perc_settings['input_files']['psm_type'] # str. PSM or search engine type, e.g. pfind, alphapept, maxquant
+#     perc_settings['input_files']['ms_file_type'] # str. Could be alphapept_hdf, thermo, ...
+#     perc_settings['input_files']['ms_files'] # list of str. MS file list to match MS2 peaks
+#     ```
 
-    Raises
-    ------
-    Exception
-        Any kinds of exception if the pipeline fails.
-    """
-    try:
-        perc_settings = global_settings['percolator']
-        output_folder = os.path.expanduser(
-            perc_settings['output_folder']
-        )
-        if not os.path.exists(output_folder):
-            os.makedirs(output_folder)
-        set_logger(
-            log_file_name=os.path.join(output_folder, 'peptdeep_rescore.log'),
-            log_level=global_settings['log_level'],
-            overwrite=True, stream=True, 
-        )
-        show_platform_info()
-        show_python_info()
+#     Raises
+#     ------
+#     Exception
+#         Any kinds of exception if the pipeline fails.
+#     """
+#     try:
+#         perc_settings = global_settings['percolator']
+#         output_folder = os.path.expanduser(
+#             perc_settings['output_folder']
+#         )
+#         if not os.path.exists(output_folder):
+#             os.makedirs(output_folder)
+#         set_logger(
+#             log_file_name=os.path.join(output_folder, 'peptdeep_rescore.log'),
+#             log_level=global_settings['log_level'],
+#             overwrite=True, stream=True, 
+#         )
+#         show_platform_info()
+#         show_python_info()
 
-        model_mgr:ModelManager = ModelManager()
-        model_mgr.reset_by_global_settings()
-        percolator = Percolator(model_mgr=model_mgr)
-        psm_df = percolator.load_psms(
-            perc_settings['input_files']['psm_files'],
-            perc_settings['input_files']['psm_type']
-        )
+#         model_mgr:ModelManager = ModelManager()
+#         model_mgr.reset_by_global_settings()
+#         percolator = Percolator(model_mgr=model_mgr)
+#         psm_df = percolator.load_psms(
+#             perc_settings['input_files']['psm_files'],
+#             perc_settings['input_files']['psm_type']
+#         )
 
-        ms_file_dict = parse_ms_file_names_to_dict(
-            perc_settings['input_files']['ms_files']
-        )
+#         ms_file_dict = parse_ms_file_names_to_dict(
+#             perc_settings['input_files']['ms_files']
+#         )
 
-        psm_df = percolator.extract_features(
-            psm_df, ms_file_dict, 
-            perc_settings['input_files']['ms_file_type']
-        )
+#         psm_df = percolator.extract_features(
+#             psm_df, ms_file_dict, 
+#             perc_settings['input_files']['ms_file_type']
+#         )
 
-        df_fdr = psm_df[
-            (psm_df.fdr<0.01)&(psm_df.decoy==0)
-        ]
-        df_fdr.to_csv(
-            os.path.join(output_folder, 'peptdeep_fdr.tsv'), 
-            sep='\t', index=False
-        )
+#         df_fdr = psm_df[
+#             (psm_df.fdr<0.01)&(psm_df.decoy==0)
+#         ]
+#         df_fdr.to_csv(
+#             os.path.join(output_folder, 'peptdeep_fdr.tsv'), 
+#             sep='\t', index=False
+#         )
         
-        psm_df = percolator.re_score(psm_df)
-        psm_df.to_csv(
-            os.path.join(output_folder, 'peptdeep.tsv'), 
-            sep='\t', index=False
-        )
-    except Exception as e:
-        logging.error(traceback.format_exc())
-        raise e
+#         psm_df = percolator.re_score(psm_df)
+#         psm_df.to_csv(
+#             os.path.join(output_folder, 'peptdeep.tsv'), 
+#             sep='\t', index=False
+#         )
+#     except Exception as e:
+#         logging.error(traceback.format_exc())
+#         raise e