+++ peptdeep/rescore/__init__.py
@@ -1,3 +1,3 @@
 from peptdeep.rescore import (
-    percolator, fdr, feature_extractor
+    feature_extractor
 )