+++ peptdeep/spec_lib/__init__.py
@@ -1,3 +1,0 @@
-from peptdeep.spec_lib import (
-    predict_lib, translate, library_factory
-)